refactor sonar 2.0
