--- actions/sonar/2.0/internal/package_manager.go
@@ -1,5 +1,0 @@
-package main
-
-type PackageManager interface {
-	Analysis(cfg *Conf) (*ResultMetas, error)
-}
