+++ actions/sonar/2.0/spec.yml
@@ -26,9 +26,9 @@ params:
   - name: sonar_password
     required: false
     desc: ${{ i18n.params.sonar_password.desc }}
-  - name: project_key
+  - name: sonar_project_key
     required: false
-    desc: ${{ i18n.params.project_key.desc }}
+    desc: ${{ i18n.params.sonar_project_key.desc }}
 accessibleAPIs:
   - path: /api/applications/<applicationId>
     method: GET
@@ -38,7 +38,7 @@ locale:
     displayName: Sonar 代码质量检测
     params.code_dir.desc: 执行代码质量分析的目录
     params.must_gate_status_ok.desc: 是否要求代码门禁检查状态为 ok
-    params.project_key.desc: 对应 sonar 里的 project key。若不填写，使用应用配置里的sonar project key
+    params.sonar_project_key.desc: 对应 sonar 里的 project key。若不填写，使用应用配置里的sonar project key
     params.sonar_host_url.desc: sonar 服务器地址，用户可以手动指定。若不填写，使用应用配置里的sonar host
     params.sonar_login.desc: 用于登录 sonar 服务器的用户名或 token。若不填写，使用应用配置里的sonar token
     params.sonar_password.desc: 用于登录 sonar 服务器的密码。若不填写，使用平台 sonar 服务自动生成的 token。特别的，当 sonar_login 为 token 而非 username 方式时，sonar_password 无需填写
@@ -47,7 +47,7 @@ locale:
     displayName: Sonar Code Quality Check
     params.code_dir.desc: The directory of code to be analyzed
     params.must_gate_status_ok.desc: Whether the code gate status is required as ok
-    params.project_key.desc: The project key in sonar. If not filled in, the system will use the project key in the application configuration
+    params.sonar_project_key.desc: The project key in sonar. If not filled in, the system will use the project key in the application configuration
     params.sonar_host_url.desc: The address of the sonar server. If not filled in, the system will use the application setting's sonar host
     params.sonar_login.desc: The username or token to login to the sonar server. If not filled in, the system will use the application setting's sonar token
     params.sonar_password.desc: "The password to login to the sonar server. If not filled in, the system will automatically generate a token. Special note: When sonar_login is token and not username, sonar_password is not required"
